+++ p2panda-spaces/src/encryption/dgm.rs
@@ -14,10 +14,10 @@ use crate::types::{ActorId, OperationId};
 /// Placeholder for DGM implementation which satisfies required trait interfaces in
 /// p2panda-encryption. Most methods perform no actual actions as group management is handled by
 /// p2panda-auth.
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct EncryptionGroupMembership {}
 
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct EncryptionMembershipState {
     pub(crate) members: HashSet<ActorId>,
 }