+++ p2panda/src/lib.rs
@@ -252,14 +252,14 @@
 //! [local-first]: https://www.inkandswitch.com/local-first-software/
 //! [p2panda]: https://p2panda.org
 mod builder;
+pub mod credentials;
 mod forge;
 pub mod network;
 pub mod node;
 pub mod operation;
 pub mod processor;
+pub mod spaces;
 pub mod streams;
-#[cfg(any(test, feature = "test_utils"))]
-pub mod test_utils;
 
 // Useful external types we want to re-export for convenience.
 #[doc(no_inline)]
@@ -270,6 +270,7 @@ pub use p2panda_net::iroh_endpoint::{EndpointAddr, RelayUrl};
 pub use p2panda_net::{NetworkId, NodeId};
 
 pub use builder::NodeBuilder;
+pub use credentials::Credentials;
 #[doc(inline)]
 pub use node::Node;