+++ p2panda/tests/api.rs
@@ -4,16 +4,16 @@ use std::time::Duration;
 
 use futures_util::StreamExt;
 use mock_instant::thread_local::MockClock;
+use p2panda::Credentials;
 use p2panda::node::AckPolicy;
 use p2panda::operation::{Extensions, LogId, Operation};
 use p2panda::streams::{
     EphemeralMessage, ProcessedOperation, StreamEvent, StreamFrom, SystemEvent,
 };
-use p2panda::test_utils::setup_logging;
 use p2panda_core::cbor::encode_cbor;
 use p2panda_core::logs::LogHeights;
-use p2panda_core::test_utils::TestLog;
-use p2panda_core::{Cursor, Hash, SigningKey, Topic};
+use p2panda_core::test_utils::{TestLog, setup_logging};
+use p2panda_core::{Cursor, Hash, Topic};
 use p2panda_net::discovery::DiscoveryEvent;
 use p2panda_store::logs::LogStore;
 use tokio::task::JoinHandle;
@@ -46,7 +46,7 @@ async fn build_and_spawn() -> Result<(), Box<dyn std::error::Error>> {
     // Customizable "builder" setup flow.
     let _node = p2panda::builder()
         .database_url("sqlite::memory:")
-        .signing_key(SigningKey::generate())
+        .credentials(Credentials::generate())
         .spawn()
         .await?;